lecture_extraction_system/requirements.txt
streamlit
openai-whisper
opencv-python
easyocr
sqlalchemy
pillow
numpy
pandas
requests
ffmpeg-python